+++ src/test/groovy/all/shared/gradle/quality/code/BaseStyleConfigWrapperPluginTest.groovy
@@ -2,6 +2,7 @@
 //  Licensed under the MIT License (MIT), see LICENSE.txt
 package all.shared.gradle.quality.code
 
+import all.shared.gradle.quality.code.complement.BaseStyleConfigWrapperComplementAction
 import all.shared.gradle.testfixtures.SpyProjectFactory
 
 import groovy.transform.CompileStatic
@@ -13,8 +14,12 @@ import org.junit.jupiter.api.Test
 import static org.junit.jupiter.api.Assertions.assertFalse
 import static org.junit.jupiter.api.Assertions.assertNotNull
 
+import static org.mockito.ArgumentMatchers.any
+import static org.mockito.Mockito.never
+import static org.mockito.Mockito.verify
+
 @CompileStatic
-class BaseStyleConfigWrapperPluginTest {
+final class BaseStyleConfigWrapperPluginTest {
   private final Project testProject = SpyProjectFactory.build()
 
   @Test
@@ -34,5 +39,26 @@ class BaseStyleConfigWrapperPluginTest {
     plugin.apply(testProject)
 
     assertFalse(testProject.properties[BaseStyleConfigWrapperPlugin.EXTENSION_NAME] instanceof BaseStyleConfigWrapperExtension)
+  }
+
+  @Test
+  void shouldAddProjectAfterEvaluateAction() {
+    final BaseStyleConfigWrapperPlugin plugin = new BaseStyleConfigWrapperPlugin()
+
+    plugin.apply(testProject)
+
+    verify(testProject)
+      .afterEvaluate(any(BaseStyleConfigWrapperComplementAction))
+  }
+
+  @Test
+  void shouldNotAddProjectAfterEvaluateActionWhenExtensionAlreadyDefined() {
+    testProject.extensions.add(BaseStyleConfigWrapperPlugin.EXTENSION_NAME, 'someValue')
+    final BaseStyleConfigWrapperPlugin plugin = new BaseStyleConfigWrapperPlugin()
+
+    plugin.apply(testProject)
+
+    verify(testProject, never())
+      .afterEvaluate(any(BaseStyleConfigWrapperComplementAction))
   }
 }